requirements.txt
pymongo
mysqlclient
mysql-connector-python
pymysql
python-dotenv
pandas